requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-groq==0.3.6
openai==1.97.1
groq==0.30.0
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1
pydantic==2.11.7
requests==2.32.4
streamlit==1.47.1
pypdf==5.9.0
python-docx==1.2.0

